requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.5
redis==4.5.1
gunicorn==20.1.0
spacy==3.5.0
pdfplumber==0.7.6
python-dotenv==1.0.0
Werkzeug==2.2.3

# Download the spaCy model in a separate step in Heroku
# python -m spacy download en_core_web_sm